MSVS 2008 solution updated: project async_connect added. Integer cast warning for x64 build fixed (async_connect).

+++ include/ma/async_connect.hpp
@@ -18,6 +18,7 @@
 
 #include <cstddef>
 #include <boost/asio.hpp>
+#include <boost/numeric/conversion/cast.hpp>
 #include <ma/config.hpp>
 #include <ma/bind_asio_handler.hpp>
 #include <ma/handler_alloc_helpers.hpp>
@@ -164,7 +165,7 @@ void async_connect(Socket& socket,
 #if defined(WIN32)
 
 #if (_WIN32_WINNT < 0x0501)
-#error The build environment doesn't support necessary Windows SDK header.\
+#error The build environment does not support necessary Windows SDK header.\
   Value of _WIN32_WINNT macro must be >= 0x0501.
 #endif
 
@@ -249,7 +250,8 @@ void async_connect(Socket& socket,
   
   // Initiate the ConnectEx operation.
   BOOL ok = connect_ex_func(native_socket, peer_endpoint.data(), 
-      peer_endpoint.size(), NULL, 0, NULL, overlapped.get());
+      boost::numeric_cast<int>(peer_endpoint.size()), NULL, 0, NULL, 
+      overlapped.get());
   DWORD last_error = ::GetLastError();
 
   // Check if the operation completed immediately.
